refactor: 머지 후 리팩토링
