--- ireon/src/docs/overview.adoc
@@ -1,39 +1,0 @@
-[[overview]]
-== Overview
-
-[[overview-context-path]]
-=== Context Path
-모든 요청 URL 은 '/api' 로 시작합니다.
-
-ex) POST '/api/member/join'
-
-[[overview-http-status-codes]]
-=== HTTP status codes
-
-|===
-| 상태 코드 | 설명
-
-| `200 OK`
-| 성공
-
-| `201 Created`
-| 등록
-
-| `302 Found`
-| 리다이랙션 상태
-
-| `400 Bad Request`
-| 잘못된 요청
-
-| `401 Unauthorized`
-| 비인증 상태
-
-| `403 Forbidden`
-| 권한 거부
-
-| `404 Not Found`
-| 존재하지 않는 요청 리소스
-
-| `500 Internal Server Error`
-| 서버 에러
-|===
