+++ ireon/src/main/java/com/uhpoo/ireon/domain/common/animal/AnimalType.java
@@ -1,4 +1,4 @@
-package com.uhpoo.ireon.domain.abandon;
+package com.uhpoo.ireon.domain.common.animal;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
