+++ ireon/src/test/java/com/uhpoo/ireon/docs/AbandonControllerDocsTest.java
@@ -9,8 +9,8 @@ import com.uhpoo.ireon.api.service.abandon.AbandonQueryService;
 import com.uhpoo.ireon.api.service.abandon.AbandonService;
 import com.uhpoo.ireon.api.service.abandon.dto.CreateAbandonDto;
 import com.uhpoo.ireon.domain.abandon.AbandonStatus;
-import com.uhpoo.ireon.domain.abandon.AnimalType;
-import com.uhpoo.ireon.domain.abandon.Gender;
+import com.uhpoo.ireon.domain.common.animal.AnimalType;
+import com.uhpoo.ireon.domain.common.animal.Gender;
 import com.uhpoo.ireon.domain.abandon.VaccinationStatus;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
